Add conditional NoResourceFound handler
